Modularise, hit field to StepInfo, O2HitIntercept
Extract StepLogger and FieldLogger from MCStepLoggerImpl.cxx in order to
allow for naturally use and integrate these objects in external code.

The hit field to StepInfo is added to flag whether a step has produced a
hit. This could be extended to contain more info of the hit like
deposited energy, spatial coordinates etc.

A interception plugin is added to intercept O2 simulation whenever
o2::data::Stack::addHit(int) is called and attach the detector ID where
the hit was registered.
